requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via django
azure-functions==1.11.2
    # via -r requirements.in
django==4.0.4
    # via -r requirements.in
sqlparse==0.4.2
    # via django
tzdata==2022.1
    # via django
whitenoise==6.1.0
    # via -r requirements.in
